+++ Tools/CirclrAuditionWorker/main.swift
@@ -1,0 +1,3 @@
+import CirclrAudio
+
+AuditionWorkerService.run(arguments:Array(CommandLine.arguments.dropFirst()))